+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixes
 
+- Exclude Sentry Web Replay by default, reducing the code in 130KB. ([#4006](https://github.com/getsentry/sentry-react-native/pull/4006))
 - `Sentry.captureMessage` stack trace is in `event.exception` (moved from `event.threads`) ([#3635](https://github.com/getsentry/sentry-react-native/pull/3635), [#3988](https://github.com/getsentry/sentry-react-native/pull/3988))
   - To revert to the old behavior (causing the stack to be unsymbolicated) use `useThreadsForMessageStack` option