+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## Unreleased
+
+### Dependencies
+
+- Bump JavaScript SDK from v7.117.0 to v8.25.0 ([#3895](https://github.com/getsentry/sentry-react-native/pull/3895))
+  - [changelog](https://github.com/getsentry/sentry-javascript/blob/develop/CHANGELOG.md#8250)
+  - [diff](https://github.com/getsentry/sentry-javascript/compare/7.117.0...8.25.0)
+
 ## 5.28.0
 
 ### Fixes